Move events to a standalone namespace
