+++ src/app/Models/Events/AiurEvent.ts
