+++ src/app/Models/Events/DissolveEvent.ts
