+++ src/app/Models/Events/FriendAcceptedEvent.ts
@@ -1,5 +1,5 @@
 import { AiurEvent } from './AiurEvent';
-import { KahlaUser } from './KahlaUser';
+import { KahlaUser } from '../KahlaUser';
 
 export class FriendAcceptedEvent extends AiurEvent {
     public target: KahlaUser;
