+++ src/app/Models/Events/NewFriendRequestEvent.ts
@@ -1,5 +1,5 @@
 import { AiurEvent } from './AiurEvent';
-import { Request } from './Request';
+import { Request } from '../Request';
 
 export class NewFriendRequestEvent extends AiurEvent {
     public request: Request;
