+++ src/app/Models/Events/NewMemberEvent.ts
@@ -1,5 +1,5 @@
 import { AiurEvent } from './AiurEvent';
-import { KahlaUser } from './KahlaUser';
+import { KahlaUser } from '../KahlaUser';
 
 export class NewMemberEvent extends AiurEvent {
     public newMember: KahlaUser;
