+++ src/app/Models/Events/SomeoneLeftEvent.ts
@@ -1,5 +1,5 @@
 import { AiurEvent } from './AiurEvent';
-import { KahlaUser } from './KahlaUser';
+import { KahlaUser } from '../KahlaUser';
 
 export class SomeoneLeftEvent extends AiurEvent {
     public leftUser: KahlaUser;
