+++ src/app/Models/Events/TimerUpdatedEvent.ts
