+++ src/app/Models/Events/WereDeletedEvent.ts
@@ -1,5 +1,5 @@
 import { AiurEvent } from './AiurEvent';
-import { KahlaUser } from './KahlaUser';
+import { KahlaUser } from '../KahlaUser';
 
 export class WereDeletedEvent extends AiurEvent {
     public trigger: KahlaUser;
