+++ src/app/Services/MessageService.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@angular/core';
-import { EventType } from '../Models/EventType';
-import { AiurEvent } from '../Models/AiurEvent';
+import { EventType } from '../Models/Events/EventType';
+import { AiurEvent } from '../Models/Events/AiurEvent';
 import Swal from 'sweetalert2';
-import { NewMessageEvent } from '../Models/NewMessageEvent';
+import { NewMessageEvent } from '../Models/Events/NewMessageEvent';
 import { Conversation } from '../Models/Conversation';
 import { Message } from '../Models/Message';
 import { ConversationApiService } from './ConversationApiService';
@@ -15,17 +15,17 @@ import * as he from 'he';
 import Autolinker from 'autolinker';
 import { Values } from '../values';
 import { ElectronService } from 'ngx-electron';
-import { TimerUpdatedEvent } from '../Models/TimerUpdatedEvent';
+import { TimerUpdatedEvent } from '../Models/Events/TimerUpdatedEvent';
 import { TimerService } from './TimerService';
-import { WereDeletedEvent } from '../Models/WereDeletedEvent';
-import { NewFriendRequestEvent } from '../Models/NewFriendRequestEvent';
-import { FriendAcceptedEvent } from '../Models/FriendAcceptedEvent';
+import { WereDeletedEvent } from '../Models/Events/WereDeletedEvent';
+import { NewFriendRequestEvent } from '../Models/Events/NewFriendRequestEvent';
+import { FriendAcceptedEvent } from '../Models/Events/FriendAcceptedEvent';
 import { Router } from '@angular/router';
 import { UserGroupRelation } from '../Models/UserGroupRelation';
-import { SomeoneLeftEvent } from '../Models/SomeoneLeftEvent';
-import { NewMemberEvent } from '../Models/NewMemberEvent';
+import { SomeoneLeftEvent } from '../Models/Events/SomeoneLeftEvent';
+import { NewMemberEvent } from '../Models/Events/NewMemberEvent';
 import { GroupConversation } from '../Models/GroupConversation';
-import { DissolveEvent } from '../Models/DissolveEvent';
+import { DissolveEvent } from '../Models/Events/DissolveEvent';
 import { HomeService } from './HomeService';
 import { GroupsApiService } from './GroupsApiService';
 import { FriendsApiService } from './FriendsApiService';
